Okay, here it is. The ngl_xxx_wrap functions still return a struct (not a pointer to a struct), but the members of the struct are now pointers. (Fred, let me know if you think the functions should be returning pointers to structures instead.)
Here's what the struct looks like:

typedef struct {
  int   *base;
  int   *contour;
  int   *vector;
  int   *streamline;
  int   *map;
  int   *xy;
  int   *xydspec;
  int   *text;
  int   *primitive;
  int   *cafield;
  int   *sffield;
  int   *vffield;

  int   nbase;
  int   ncontour;
  int   nvector;
  int   nstreamline;
  int   nmap;
  int   nxy;
  int   nxydspec;
  int   ntext;
  int   nprimitive;
  int   ncafield;
  int   nsffield;
  int   nvffield;
} nglPlotId;


I still can't get all of the test plots to maximize properly if
the output is PDF or PS. This is because if you try to maximize
a plot when something has already been drawn to the screen, then all
bets are off.

Also, I still sometimes see the warning messages:

  warning:PSWorkstationInitialize:Device X Coordinates invalid, defaulting

I think this has something to do with some *.o not getting rebuilt, even
though a *.h file it's including has been updated. I'll have to do
a build from scratch to get rid of this, probably.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@7250 10fe2890-8a40-4696-8b1d-7ccbb19c398e