Removed a duplicate resource setting.
Made "nglSpreadColors" default to True. Fred, did we decide on this?
By default, the colors for a filled contour or vector plot will *not*
span the full colormap in GSUN, but most people seem to want the
colors to span, so I thought we should make it the default for PyNGL.

Comments?


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@7260 10fe2890-8a40-4696-8b1d-7ccbb19c398e