Mary,
This PyNGL update is for testing and evaluation purposes only.
Please review what I have done, as described below, and tell me
if this serves your purposes.  If so, we can continue modifying
all appropriate functions.

In gsun.h:
---------

 1.)  I have added a structure:

        /*
         *  Structure for passing resource names.
         */
        typedef struct _res_names {
          int number;
          char **strings;
        } res_names;

 2.)  In the function prototype for ngl_open_wks_wrap I have
      added an argument of type "res_names *" as the final argument:

   extern int ngl_open_wks_wrap(const char *, const char *, int, res_names *);


In gsun.c:
---------

 1.)  I added a final argument to ngl_open_wks_wrap:

         int ngl_open_wks_wrap(const char *type, const char *name,
                               int wk_rlist, res_names *rn)

 2.)  For purposes of illustration I have put some print statements
      in ngl_open_wks_wrap to print out the structure members of the
      resource names structure:


  /*********************************************************************
   *  Print the contents of the resource name structure.
   *********************************************************************/
    printf ("\nFrom ngl_open_wks_wrap, number of resources = %d\n",rn->number);
    for (len = 0; len < rn->number; len++) {
      printf("    string %3d = %s\n",len,(rn->strings)[len]);
    }
    printf("\n");
  /*********************************************************************
   *  End print
   *********************************************************************/


Simple test
-----------

I have written a trivial Python test code:

  import Numeric
  from Ngl import *

  cmap = Numeric.zeros((2,3),Numeric.Float0)
  cmap[0] = [1.,1.,1.]
  cmap[1] = [0.,0.,0.]

  rlist = Resources()
  rlist.wkColorMap = cmap
  rlist.wkBull = 55.
  rlist.wkNonsense = "string_0"
  wks = ngl_open_wks("ncgm","ngl05p",rlist)

  ngl_end()

to invoke the print statements in ngl_open_wks_wrap.

The output from the printf statements in ngl_open_wks_wrap should be:

 From ngl_open_wks_wrap, number of resources = 3
    string   0 = wkColorMap
    string   1 = wkBull
    string   2 = wkNonsense


Comment on freeing space
------------------------

I am using SWIG to insert the proper code into the C wrappers
for the resource names structure.  I dynamically allocate
space for the string pointers in the structure and I can see no way,
using SWIG, to free that space.  So, I am going to have to leave it
to you to free the strings member of the res_names structure.


Fred


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@7525 10fe2890-8a40-4696-8b1d-7ccbb19c398e