The setup.py file seems to be working, and I have finally made a decision about where to put the supplemental files:
    $python_prefix/lib/pythonx.x/site-packages/PyNGL/ncarg

I've added a skeleton cut at a setup file that does the build of
PyNGL (in lieu of using a Makefile). This file still needs a lot of
work and is just a work in progress. The setup.py file assumes the
build of "_hlu.so" has already occurred. (It just does a binary
distribution installation, not a source code installation.)


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@7745 10fe2890-8a40-4696-8b1d-7ccbb19c398e