Some new plot defaults were added, mostly having to do with the labelbar. These are plot defaults that are different than what NCL uses.  Since there were several labelbar resources, I made the setting of them their own subroutine, called "set_labelbar_res".
Here's the full list of resources added:

  1. pmTickMarkDisplayMode = "Always" for map projections
  2. lbPerimOn             = False
  3. lbLabelAutoStride     = True
  4. lbAutoManage          = False if lbLabelFontHeightF is set
  5. pmLabelBarSide        = "Bottom" if lbOrientation = "Horizontal" and
     pmLabelBarSide        = "Right" if lbOrientation = "Vertical".

Also, added a few more cases where the labelbar gets turned on, like
in the vector routines when vcMonoLineArrow is set to False.

Added a function called "check_res_value" that allows you to check the
value of a resource that can be of type string or integer. For
example, let's say you want to check if the user has set the
"lbOrientation" resource to "horizontal" so that you can set
"pmLabelBarSide" to "Bottom". Since this resource can also be set with
the integer 0, you can do this with:

  if(check_res_value(rlist["lbOrientation"],"Horizontal",0)):
    rlist["pmLabelBarSide"] = "Bottom"

I tried to convert all instances in Ngl.py to use this new
function where appropriate.

--Mary


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@7823 10fe2890-8a40-4696-8b1d-7ccbb19c398e