Ngl.py: ------ Created routines "set_vector_res" and "set_map_res" that are used to set the vector and map resources that we want to be the default for PyNGL.
Also started setting some of the cnMonoXXX resources to False if
their corresponding plural resources (i.e. cnLineDashPatterns)
are being set.

In "set_spc_res", removed the hard-coded indexing of the
"resource_name" variable. This was causing a bug where an unrecognized
resource (nglSpreadColorsEnd) slipped through the cracks because it
was a super-set of a correctly spelled resource (nglSpreadColors).

I need to do testing on this last change, to make sure I didn't
break something else.

In setup.py, it is now easier for the user to switch between a binary
and a source installation. The source installation assume that
you have just built all the necessary files that PyNGL depends
on, and installed them somewhere.

I wouldn't be up this late except that I had to respond to yet another
email from my dad that contained some article on why Bush needs to be
reelected. <sigh>


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@7858 10fe2890-8a40-4696-8b1d-7ccbb19c398e