gsun.h: Removed _res_names structure because it doesn't seem to be used by anybody.
Removed a call to NhlClear on the app object, because this was
clearing any resources set by the hlu_wrap.c code.

With this check-in, users can now have a resource filename prefix that is
different from their output filename prefix.  They can also indicate
a different directory for the resource file.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@7916 10fe2890-8a40-4696-8b1d-7ccbb19c398e