Now that we have the capability to overlay a contour, streamline, or vector plot on an irregular plot (in order to force a linear or log axis), I added some tests for tickmark resources, so that they get applied to the IrregularPlot class.  This required creating a new resource list in Ngl.py that retains just the "tm" and "pmTick" resources and passes them to contour_wrap, streamline_wrap, vector_wrap, and vector_scalar_wrap.
In scale_plot, I removed all the code for retrieving the tickmark
lengths and label font heights (done to eventually make them the same
size on both axes), and replaced it with the much easier
"tmEqualizeXySizes = True" resource.

Changed the retrieval of the resources trXCoordPoints and trYCoordPoints
to be a multi-dimensional array (instead of a 1D array). This handles
the case of the arrays being 2D.

There is still a small problem with how to check if tickmarks have
been turned off in a draw plot (like a ContourPlot). It seems that
even though no tickmarks are appearing, "pmTickMarkDisplayMode"
is still returning "ALWAYS".  Dave B is going to look into this.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@7996 10fe2890-8a40-4696-8b1d-7ccbb19c398e