Another fix to handle problems with overlaying contours, vectors, and streamlines on an irregular plot in order to linearize or logize one of the axes.
The code to do all this is turning into a real pain in the *ss, and I'm
about to give up on it and make the user do all the work.  :-)

The problem is trying to decide what resources that the user has
applied to the contour plot that should also apply to the irregular
plot. For example, setting the viewport width and height is something
you want to apply to the irregular plot class.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@8025 10fe2890-8a40-4696-8b1d-7ccbb19c398e