In testing the Ngl.get_xxxx functions, I came to the sinking realization that I wasn't returning the DataSpec object that is created on one's behalf when an XY plot object is created. This DataSpec object is how resources that affect the XY plot (like xyLineColors, xyMarkers, xyDashPatterns, etc) are set.
The problem is that if a user tries to use the XY plot object
to retrieve any of these resources, they won't get back the
expected values, because the XY plot object doesn't know about
these resources.

So, I am returning the DataSpec object as part of the XY PlotId
structure (the variable for this was already in the PlotId
structure). Users will have to use this in order to retrieve
these kind of resources (I will create an example shortly).

I suppose one could add code to all of the get_xxxx resource functions
to check if the user is passing an XY plot object, and whether they
are trying to retrieve one of the XY DataSpec resources. If so,
then the get_xxxx function could smartly realize what the user is doing,
and do the right thing by invoking the DataSpec object instead.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@8589 10fe2890-8a40-4696-8b1d-7ccbb19c398e