This is a pretty significant change. I've added a non-empty sysresfile that sets some global resources for all things PyNGL:
   1. The background/foreground colors are set to white/black.
   2. *Font is set to helvetica
   3. The TextFunc code is set to '~'.

Note that I didn't set a different colormap. We can do this, but it may
significantly change the looks of some of the graphical examples.

As it is, the above change is probably going to affect most if not all
the examples, mostly in the area of the font being helvetica.

Also, there may be some tutorial examples that will need to be updated (or
at least the explanation) to explain what's going on here.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@8603 10fe2890-8a40-4696-8b1d-7ccbb19c398e