Added an example that draws the HSV color wedges and also tests the six color conversion functions.
I want to add another example of color boxes and I need to update
the documentation for the color conversion functions to document
that they are called like:

  r,g,b=Ngl.hsvrgb(h,s,v)

rather than the way they are documeneted.

Also, I need to change the sources for the conversion functions to
accept arrays as well as scalars.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@8623 10fe2890-8a40-4696-8b1d-7ccbb19c398e