Removed a couple of min and max calculations on arrays that were causing errors with NumPy and were never used anyway.
I now have a version of Ngl and versions of all of the examples
that run with NumPy (together with Dave B.'s version of Nio that
runs with NumPy).  This is not to say that all NumPy considerations
have been shaken out.  For example, I am not sure about returning
scalars - should they be NumPy scalars or Python scalars?


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@8840 10fe2890-8a40-4696-8b1d-7ccbb19c398e