Here's the setup.py file. To build PyNGL, use:
  python setup.py build_ext -L/sw/lib

if you are on a Mac.  It will look for the pynglex examples in ../examples.
If this director doesn't exist, or if you want to use CVS to check out
the examples, then setenv USE_CVS to 1 before you run the above.

Then, to install everything:

   python setup.py install

To install PyNIO, you will have to do this separately.

--Mary


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@8929 10fe2890-8a40-4696-8b1d-7ccbb19c398e