Made changes to setup.py so it now checks for your platform, and tries to do the right thing. So, for the Mac, you shouldn't need to type:
   python setup.py build_ext -L/sw/lib

any longer.  The following should suffice:

   python setup.py build_ext

Also, I implemented Fred's suggestion of only including the *py and
*.res files from the ../examples directory when trying to gather
up all the pynglex files.

On tempest, the compilation is a mess. The setup.py file is
automatically including "-all" on the load line, and this causes all
symbols from all archived libraries to be exported, which in turn
causes an undefined reference. I don't see in the distutils
documentation how to turn off "-all".  Even adding the additional
"-notall" argument doesn't work, because Python's distuils puts it at
the end of the load line, and it needs to be before the loading of the
*.o files.

So, I had to create a "build_on_irix64" script for now until we can
get this resolved. To build on tempest, use this script, and not
"python setup.py install".


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@8953 10fe2890-8a40-4696-8b1d-7ccbb19c398e