This check-in changes the way the __array_module__ and __array_module_version__ attributes are created. Rather than being created on the fly when Ngl is imported, they are created at the time "python setup.py install" is executed. This way, you can "stamp" that version of Ngl with the version of Numeric or NumPy that it was built with.
The setup.py script constructs the pyngl_version.py file from scratch, so
that's why I'm removing it from CVS. It uses the "version" file to get
the PyNGL version, and then it puts the version, the two above attributes,
and the value of HAS_NUM in the file.

The Ngl.py script, then, imports pyngl_version and gets all this information
and does the appropriate thing with Numeric or NumPy.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@8983 10fe2890-8a40-4696-8b1d-7ccbb19c398e