Okay, here it is, all the mods to get two separate packages built for  Numeric and numpy.
If no environment variable is set, then a Numeric version of PyNGL
will be built and installed to the package PyNGL.

If USE_NUMPY is set then a numpy version of PyNGL will be built
and installed to the package PyNGL.

If USE_NUMERPY is set (in which case USE_NUMPY will be ignored),
then both a Numeric and numpy version of PyNGL will be built, and
installed to packages PyNGL and PyNGL_numpy respectively.

For now, I will be setting USE_NUMERPY. The other two cases are for
when we can safely move completely over to numpy.

I've made similar changes to PyNIO's setup.py, which I will check in
shortly.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9003 10fe2890-8a40-4696-8b1d-7ccbb19c398e