dded a new function, Ngl.change_coord_limits, that allows you to add a new min/max limit for the X and/or Y coordinate arrays. It creates a new data array with the new size and returns the new coord array(s).
This function is useful if you want to set new axes limits for
a plot that is using coordinate arrays. You can't do this with
the usual way of using trXMinF, trXMaxF, trYMinF, or trYMaxF,
so this function allows you to get around that.

See "irregular.py" (after I check it in) for an example of how to
use this function.

Also fixed Ngl.merge_colormaps so that it will work with numpy (was
referencing typecode).

Changed mod_pynglex_files.py so that it would change "import MA"
to "import numpy.core.ma as MA" for the multi_plot.py example.

Updated the setup.build.py script so that it also removes the *.so
files (it was already doing the *.o files) for the case of
USE_NUMERPY being set.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9292 10fe2890-8a40-4696-8b1d-7ccbb19c398e