Added the setup.py file needed to build the fplib.so module and install it to the PAF packages directory.  You still do an "import NglA" to get access to the functions.
Eventually, this will all become part of the PyNGL package. For now,
for test purposes, it is stand-alone.

To install PAF, type:

   python setup.py install

Note: if you have an old "fplib.so" in this directory, this is the
one that will get loaded, probably. You should remove this before
doing anything else.

Changed fplibmodule.c to include "Error.h" as "<ncarg/hlu/Error.h>".
Added some version stuff to NglA.py.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9392 10fe2890-8a40-4696-8b1d-7ccbb19c398e