Changed the interface to linmsg as per:
  def linmsg(x, end_pts=0, max_msg="MAX", fill_value=1.e20):

The end points are now treated depending on whether end_pts is
greater or equal to zero, or less than zero.  The maximum allowable
consecutive missing values is the size of the rightmost dimension
of x by default (max_msg="MAX"), or the interger value of max_msg if
that argument is specified as an integer.

Is this any better?


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9396 10fe2890-8a40-4696-8b1d-7ccbb19c398e