I added a bunch of little logic functions, like is_numpy, is_numeric, is_numpy_scalar, is_numeric_scalar, is_python_scalar, and is_scalar.
I also added a promote_scalar routine, as Fred and I discussed. This
routine converts arguments to numeric or numpy arrays of at least 1
dimension. This is to avoid the problem of getting an input argument
that has an ndims of 0, and hence potentially causing the C extension
module to fail.

Whether to promote to a numeric or numpy array is determined first via
the type of fplib module being loaded, and if that fails, then the
other type is used.

I then modified the "chiinv" interface to utilize these functions.

The return argument of chiinv is now based on what types of arguments
were passed into the function. If either of the input arguments are
numpy arrays, then a numpy array is returned. Otherwise, if any of the
input arguments are numeric arrays, then a numeric array is returned.
If neither of the arguments are numeric or numpy arrays, then the type
is determined by the type of the fplib module imported.

Fred, we can discuss this logic tomorrow to see if you agree with it.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9418 10fe2890-8a40-4696-8b1d-7ccbb19c398e