This check-in switches our PyNGL distribution to build the NumPy module by default.
Now, to build the Numeric module, you must either "setenv
USE_NUMERIC", or "setenv USE_NUMERPY" (which builds both). USE_NUMPY
is still recognized, but not needed (since it is now the default).

I have changed all the pynglex examples to be numpy-based instead of
numeric-based.  I changed "pynglex" itself so that in order to generate
the numeric examples, you have to type "pynglex --numeric ...".

Note that the examples color3, skewt1, skewt2, skewt3, and wmbarbmap
are not working because of incompatible types in a couple of routines.

I'm not sure yet if this is due to these changes, or to the last changes
to hlu_wrap.c. I may work with Fred on this, if he is willing.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9575 10fe2890-8a40-4696-8b1d-7ccbb19c398e