In order to be able to test both python 2.4 and 2.5 concurrently, it helps if you can have pynglex invoke the current version of python. So, I've done a similar thing that python itself does, and that is I've created "pynglex2.4" (for python 2.4) or "pynglex2.5" (for python 2.5).
The pynglexX.Y script specifically invokes pythonX.Y.  This way, if
you want to generate a python 2.5 version of an example, you can type:

   pynglex2.5 topo1

To generate the 2.4 version:

   pynglex2.4 topo1

Of course, you must have the 2.4 and/or 2.5 versions of PyNGL and PyNIO
installed to their respective pythonX.y/site-packages directories.

The "pynglex" file is still there and is the same. It just invokes
"python" to generate the examples.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9600 10fe2890-8a40-4696-8b1d-7ccbb19c398e