Added a fix so that when a map plot is created, the X and Y axes labels (if any) are scaled so they are the same size as each other. They shouldn't get scaled if the user is explicitly setting the font height.
Added a second fix so that if "ti" resources are set for any plot that
is overlaid on a map, the "ti" resources are applied to the map, and
not to the plot being overlaid. This is so they don't go through a
transformation and hence potentially get their font sizes resized
(even if set by user).  Also set "pmTitleDisplayMode" to "conditional"
for map plots (it is "Never" by default, which means your titles won't
show up).


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9666 10fe2890-8a40-4696-8b1d-7ccbb19c398e