This adds the first cut at a wrapper script that asks you questions about the Fortran function you want to wrap for PyNGL, and generates a *.c wrapper file for you.
Note: this is not complete. I've only tested it on a very simple
test, the "chiinv" function. It still does things a bit redundantly,
so you have to go in and remove unneeded code if you want it to
be clean.

I'm checking in a modified version of chiinvP.c This version
is from running "python wrapper.py < chiinv.txt" and then
removing some unneeded code. You can try this yourself and
then do a "cvs diff chiinvP.c"  to see what's different.

Updated the chiinv_test.py script to default to NumPy and
to have more comments.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9802 10fe2890-8a40-4696-8b1d-7ccbb19c398e