First pass at adding recognition of masked arrays to graphical functions. Right now, only Ngl.xy recognized masked arrays.
I created a function called get_arr_and_fill_value that
if it encounters a masked array, it returns a NumPy/Numeric
array, with the missing areas filled in with the fill_value,
and the fill_value itself. If it is not a masked array, then
the same array is returned, and "None" is returned for
the fill_value.

In Ngl.xy, if a masked array is encountered for either the X or Y
array, the corresponding caXMissingValueV and/or caYMissingValueV
resource is set.

We only have one pynglex example, ngl10p.py, that sets these
ca*MissingValueV resources, so I used this example as a test.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9832 10fe2890-8a40-4696-8b1d-7ccbb19c398e