After my last check-in to fix a problem with having added labelbar and legend to the nglPlotId list of objects, I realized that I hadn't done this everywhere necessary.
So, I think I've done that here, and I've written a readme file on how
to do this for new objects in the future.

Adding a new object is rather tedious, and I'm wondering if I'm doing
something wrong. The biggest part is all the changes you have to make
to hlu.i.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@10086 10fe2890-8a40-4696-8b1d-7ccbb19c398e