These are  some significant changes. I'm checking in a working version, but then plan to go back and remove recognition of the MaskedArray module, since this is becoming part of numpy 1.0.5
Here are the changes, some of them made to work with Dave's recent changes
to PyNIO:

Resource lists can now have masked arrays as values.

Fixed bug in which if one of the vcMonoXXXColor resources was being
set to True, the labelbar was not being turned on (as it should be).

Added recognition of masked arrays to primitive functions. The Ngl.text
functions are not done yet, and I need to do some more testing here.

Added tests to vector routines to make sure input arrays are 2D.

Fixed bus error problem where if a CoordArrays was bad, then it
doesn't get passed to an XyPlo objectt. This is different than how
ContourPlot, VectorPlot, and StreamlinePlot handle things, because
they allow "bad" data objects to be handed to them. In this case, you
get a "no data" message inside a TickMark object.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@10132 10fe2890-8a40-4696-8b1d-7ccbb19c398e