This example was not producing the create image with regard to the labelbar.
This was because the labelbar being created was supposed to be
associated with the map plot, and not the contour plot (which is a bit
unusual).

To fix this, I moved some of the resources inside the cn12p.res file
to inside the script. I also modified the script so that instead
of using Ngl.contour_map, it now uses Ngl.contour, Ngl.map, and
Ngl.overlay.  This allows the labelbar resources to be applied to
the map's labelbar (allowing more control over the labelbar, like
how many boxes it contains).


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@10450 10fe2890-8a40-4696-8b1d-7ccbb19c398e