This check-in completely rewrites the wrapper for "vinth2p" and hopefully fixes the memory bleed. There still seems to be a memory hit, but that's only initially, which is to be expected.
Instead of having a Python interface that does several things to the
input before calling the C interface, this new version calls the C
interface directly. This new C interface takes care of converting the
input to double as appropriate.

The one difference in this routine and the old version is that the new
version returns the type as 'numpy.float64' . The old version returned
'numpy.float32'. I'm not sure yet if I want to do things like the NCL
interface, where a double (float64) is output if and only if "datai"
is double.


git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@11027 10fe2890-8a40-4696-8b1d-7ccbb19c398e