This fixes PYNGL-27.
A user was setting a resource to force an axis to be linearized,
but the coordinates for the X axis were never set, causing a bus
error.

The fix first required a fix to NCL-718 (NULL is now returned for
sf/vf/X/YArray if they were not set by user), and also, the axis
is forced to be linear if no coordinates are set for it.




git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@11677 10fe2890-8a40-4696-8b1d-7ccbb19c398e