Fixed a bug in the nice_cntr_levels code. Removed nice_mnmxintvl.
This code was contributed by Lou Wicker, and we decided to
combine nice_mnmxintvl and nice_cntr_levels into one function.
The new function has a new flag, returnLevels, that indicates
whether the levels should be returned.

Lou and I created a test script for this function which is also being
checked in.



git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@11700 10fe2890-8a40-4696-8b1d-7ccbb19c398e