Added an "add_lat_90" function, similar to add90LatX/Y.
This addresses PYNGL-85, where an MPI user asked for this function.

It simply takes a 2D array that has latitude as one of the dimensions,
and adds two extra elements to represent lat=90 and lat=-90. The values
are set to missing in these locations.

To use it, do something like this:

  T0,lat = Ngl.add_lat_90(T[0,:,:],lat,"Y")



git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@14652 10fe2890-8a40-4696-8b1d-7ccbb19c398e